fixed Modal delete button component
