+++ components/Modal.js
@@ -1,0 +1,36 @@
+import React, { useEffect, useState } from "react";
+import ReactDOM from "react-dom";
+import styled from "styled-components";
+
+export default function Modal({ show, onClose, children }) {
+  const [isBrowser, setIsBrowser] = useState(false);
+
+  useEffect(() => {
+    setIsBrowser(true);
+  }, []);
+
+  const modalContent = show ? (
+    <StyledModalOverlay>{children}</StyledModalOverlay>
+  ) : null;
+
+  if (isBrowser) {
+    return ReactDOM.createPortal(
+      modalContent,
+      document.getElementById("modal-root")
+    );
+  } else {
+    return null;
+  }
+}
+
+const StyledModalOverlay = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.5);
+`;
